Trim fclCombineMisc 003d71d0 locals

+++ src/Yajima/y_runtime.c
@@ -5339,7 +5339,7 @@ void FUN_00434e60(void)
   
   for (iVar3 = 1; iVar3 < 4; iVar3 = iVar3 + 1) {
     if (*(int *)((u8 *)DAT_0087190c + iVar3 * 0x1c0) != 0) {
-      uVar2 = (u8)iVar3;
+      uVar2 = iVar3;
       cVar1 = FUN_0043c7f0(uVar2);
       if (cVar1 != '\x06') {
         cVar1 = FUN_0043c7f0(uVar2);
